internal/integration/prosody: extract all Lua code
This patch extracts all Lua code into external files and embeds them for
use. This will make editing the Lua easier and makes the Go source
easier to read.
