+++ internal/integration/prosody/prosody.go
@@ -173,117 +173,6 @@ func MUC(domain string, modules ...string) integration.Option {
 	return Component(domain, "", "muc", modules...)
 }
 
-// Channel configures the MUC component (if loaded) with a default channel or
-// channels.
-func Channel(domain string, c ...ChannelConfig) integration.Option {
-	const modName = "muc_defaults"
-	return func(cmd *integration.Cmd) error {
-		cfg := getConfig(cmd)
-		comp := cfg.Component[domain]
-		comp.MUCDefaults = append(comp.MUCDefaults, c...)
-		comp.Modules = append(comp.Modules, modName)
-		cfg.Component[domain] = comp
-		cmd.Config = cfg
-		return integration.TempFile("mod_"+modName+".lua", func(_ *integration.Cmd, w io.Writer) error {
-			_, err := io.WriteString(w, `local log = module._log;
-local params = module:get_option("default_mucs", {});
-local jid_bare = require "util.jid".bare;
-
-
-local function set_affiliations(room, affiliations)
-	for affiliation, jids in pairs(affiliations) do
-		for i, jid in pairs(jids) do
-			module:log("debug", "Setting affiliation %s for jid %s", affiliation, jid);
-			room:set_affiliation(true, jid_bare(jid), affiliation);
-		end
-	end
-end
-
-
-local function configure_room(room, config)
-	local should_save = false;
-	if config.name ~= nil then
-		should_save = room:set_name(config.name) or should_save;
-	end
-	if config.description ~= nil then
-		should_save = room:set_description(config.description) or should_save;
-	end
-	if config.allow_member_invites ~= nil then
-		should_save =
-			room:set_allow_member_invites(config.allow_member_invites)
-			or should_save;
-	end
-	if config.change_subject ~= nil then
-		should_save =
-			room:set_changesubject(config.change_subject)
-			or should_save;
-	end
-	if config.history_length ~= nil then
-		should_save =
-			room:set_historylength(config.history_length)
-			or should_save;
-	end
-	if config.lang ~= nil then
-		should_save = room:set_language(config.lang) or should_save;
-	end
-	if config.pass ~= nil then
-		should_save = room:set_password(config.pass) or should_save;
-	end
-	if config.members_only ~= nil then
-		should_save =
-			room:set_members_only(config.members_only)
-			or should_save;
-	end
-	if config.moderated ~= nil then
-		should_save = room:set_moderated(config.moderated) or should_save;
-	end
-	if config.persistent ~= nil then
-		should_save = room:set_persistent(config.persistent) or should_save;
-	end
-	if config.presence_broadcast ~= nil then
-		should_save = room:set_presence_broadcast(config.presence_broadcast) or should_save;
-	end
-	if config.public ~= nil then
-		should_save = room:set_hidden(not config.public) or should_save;
-	end
-	if config.public_jids ~= nil then
-		should_save =
-			room:set_whois(config.public_jids and "anyone" or "moderators")
-			or should_save;
-	end
-	if config.logging ~= room._data.logging then
-		room._data.logging = config.logging;
-		should_save = true;
-	end
-	if should_save then
-		room:save(true);
-	end
-end
-
-
-local i, room_data;
-for i, room_data in pairs(params) do
-	local host = module.host;
-	local room_jid = room_data.jid_node.."@"..host;
-	local mod_muc = prosody.hosts[host].modules.muc;
-	local room = mod_muc.get_room_from_jid(room_jid);
-	if not room then
-		module:log("debug", "Creating new room %s", room_jid);
-		-- We don't pass in the config, so that the default config is set first.
-		room = mod_muc.create_room(room_jid);
-	else
-		module:log("debug", "Configuring already existing room %s", room_jid);
-	end
-	configure_room(room, room_data.config);
-	if room_data.affiliations then
-		set_affiliations(room, room_data.affiliations);
-	end
-end`)
-			return err
-		})(cmd)
-	}
-}
-
 // Component adds an component with the given domain and secret to the config
 // file.
 // If a name is provided the component must be a builtin.
@@ -418,183 +307,6 @@ func Set(key string, v interface{}) integration.Option {
 		cfg.Options[key] = v
 		cmd.Config = cfg
 		return nil
-	}
-}
-
-// Bidi enables bidirectional S2S connections.
-func Bidi() integration.Option {
-	// TODO: Once Prosody 0.12 is out this module can be replaced with the builtin
-	// mod_s2s_bidi. See https://mellium.im/issue/78
-	const modName = "bidi"
-	return func(cmd *integration.Cmd) error {
-		err := Modules(modName)(cmd)
-		if err != nil {
-			return err
-		}
-		return integration.TempFile("mod_"+modName+".lua", func(_ *integration.Cmd, w io.Writer) error {
-			_, err := io.WriteString(w, `
--- Bidirectional Server-to-Server Connections
--- http://xmpp.org/extensions/xep-0288.html
--- Copyright (C) 2013 Kim Alvefur
---
--- This file is MIT/X11 licensed.
---
-local add_filter = require "util.filters".add_filter;
-local st = require "util.stanza";
-local jid_split = require"util.jid".prepped_split;
-local core_process_stanza = prosody.core_process_stanza;
-local traceback = debug.traceback;
-local hosts = hosts;
-local xmlns_bidi_feature = "urn:xmpp:features:bidi"
-local xmlns_bidi = "urn:xmpp:bidi";
-local secure_only = module:get_option_boolean("secure_bidi_only", true);
-local disable_bidi_for = module:get_option_set("no_bidi_with", { });
-local bidi_sessions = module:shared"sessions-cache";
-
-local function handleerr(err) log("error", "Traceback[s2s]: %s: %s", tostring(err), traceback()); end
-local function handlestanza(session, stanza)
-	if stanza.attr.xmlns == "jabber:client" then --COMPAT: Prosody pre-0.6.2 may send jabber:client
-		stanza.attr.xmlns = nil;
-	end
-	-- stanza = session.filter("stanzas/in", stanza);
-	if stanza then
-		return xpcall(function () return core_process_stanza(session, stanza) end, handleerr);
-	end
-end
-
-local function new_bidi(origin)
-	if origin.type == "s2sin" then -- then we create an "outgoing" bidirectional session
-		local conflicting_session = hosts[origin.to_host].s2sout[origin.from_host]
-		if conflicting_session then
-			conflicting_session.log("info", "We already have an outgoing connection to %s, closing it...", origin.from_host);
-			conflicting_session:close{ condition = "conflict", text = "Replaced by bidirectional stream" }
-		end
-		bidi_sessions[origin.from_host] = origin;
-		origin.is_bidi = true;
-		origin.outgoing = true;
-	elseif origin.type == "s2sout" then -- handle incoming stanzas correctly
-		local bidi_session = {
-			type = "s2sin"; direction = "incoming";
-			incoming = true;
-			is_bidi = true; orig_session = origin;
-			to_host = origin.from_host;
-			from_host = origin.to_host;
-			hosts = {};
-		}
-		origin.bidi_session = bidi_session;
-		setmetatable(bidi_session, { __index = origin });
-		module:fire_event("s2s-authenticated", { session = bidi_session, host = origin.to_host });
-		local remote_host = origin.to_host;
-		add_filter(origin, "stanzas/in", function(stanza)
-			if stanza.attr.xmlns ~= nil then return stanza end
-			local _, host = jid_split(stanza.attr.from);
-			if host ~= remote_host then return stanza end
-			handlestanza(bidi_session, stanza);
-		end, 1);
-	end
-end
-
-module:hook("route/remote", function(event)
-	local from_host, to_host, stanza = event.from_host, event.to_host, event.stanza;
-	if from_host ~= module.host then return end
-	local to_session = bidi_sessions[to_host];
-	if not to_session or to_session.type ~= "s2sin" then return end
-	if to_session.sends2s(stanza) then return true end
-end, -2);
-
--- Incoming s2s
-module:hook("s2s-stream-features", function(event)
-	local origin, features = event.origin, event.features;
-	if not origin.is_bidi and not origin.bidi_session and not origin.do_bidi
-	and not hosts[module.host].s2sout[origin.from_host]
-	and not disable_bidi_for:contains(origin.from_host)
-	and (not secure_only or (origin.cert_chain_status == "valid"
-	and origin.cert_identity_status == "valid")) then
-		if origin.incoming == true then
-			module:log("warn", "This module can now be replaced by mod_s2s_bidi which is included with Prosody");
-		end
-		module:log("debug", "Announcing support for bidirectional streams");
-		features:tag("bidi", { xmlns = xmlns_bidi_feature }):up();
-	end
-end);
-
-module:hook("stanza/urn:xmpp:bidi:bidi", function(event)
-	local origin = event.session or event.origin;
-	if not origin.is_bidi and not origin.bidi_session
-	and not disable_bidi_for:contains(origin.from_host)
-	and (not secure_only or origin.cert_chain_status == "valid"
-	and origin.cert_identity_status == "valid") then
-		module:log("debug", "%s requested bidirectional stream", origin.from_host);
-		origin.do_bidi = true;
-		return true;
-	end
-end);
-
--- Outgoing s2s
-module:hook("stanza/http://etherx.jabber.org/streams:features", function(event)
-	local origin = event.session or event.origin;
-	if not ( origin.bidi_session or origin.is_bidi or origin.do_bidi)
-	and not disable_bidi_for:contains(origin.to_host)
-	and event.stanza:get_child("bidi", xmlns_bidi_feature)
-	and (not secure_only or origin.cert_chain_status == "valid"
-	and origin.cert_identity_status == "valid") then
-		if origin.outgoing == true then
-			module:log("warn", "This module can now be replaced by mod_s2s_bidi which is included with Prosody");
-		end
-		module:log("debug", "%s supports bidirectional streams", origin.to_host);
-		origin.sends2s(st.stanza("bidi", { xmlns = xmlns_bidi }));
-		origin.do_bidi = true;
-	end
-end, 160);
-
-function enable_bidi(event)
-	local session = event.session;
-	if session.do_bidi and not ( session.is_bidi or session.bidi_session ) then
-		session.do_bidi = nil;
-		new_bidi(session);
-	end
-end
-
-module:hook("s2sin-established", enable_bidi);
-module:hook("s2sout-established", enable_bidi);
-
-function disable_bidi(event)
-	local session = event.session;
-	if session.type == "s2sin" then
-		bidi_sessions[session.from_host] = nil;
-	end
-end
-
-module:hook("s2sin-destroyed", disable_bidi);
-module:hook("s2sout-destroyed", disable_bidi);
-`)
-			return err
-		})(cmd)
-	}
-}
-
-// TrustAll configures prosody to trust all certificates presented to it without
-// any verification.
-func TrustAll() integration.Option {
-	const modName = "trustall"
-	return func(cmd *integration.Cmd) error {
-		err := Modules(modName)(cmd)
-		if err != nil {
-			return err
-		}
-		return integration.TempFile("mod_"+modName+".lua", func(_ *integration.Cmd, w io.Writer) error {
-			_, err := io.WriteString(w, `
-module:set_global();
-
-module:hook("s2s-check-certificate", function(event)
-	local session = event.session;
-	module:log("info", "implicitly trusting presented certificate");
-	session.cert_chain_status = "valid";
-	session.cert_identity_status = "valid";
-	return true;
-end);`)
-			return err
-		})(cmd)
 	}
 }
